+++ CHANGES.md
@@ -5,5 +5,11 @@ This file records the changes in *PyLf* library in reverse chronological order.
 [Gsllchb]
 * Improve docstring
 [Gsllchb]
-* **将黑色作为template的缺省颜色**
+* **将黑色作为*template*的缺省颜色**
+[Gsllchb]
+* **字体宽度从由*font_size*决定改为由每个字符自己的信息决定**
+[Gsllchb]
+* **将(lambda c: False)作为*template*的缺省*is_half_char***
+[Gsllchb]
+* **将是否在常见非开头字符集中作为*template*的缺省*is_end_char***
 [Gsllchb]